Implements the plugin API
The plugin API allows you to specify can-ssr plugins that will run in
the context of your app while server-side rendering. These are useful
for providing extra features.

For example the **canjs** plugin (included by default) provides some
stache helpers like `{{isProduction}}` and `{{asset "css"}}`.

Other frameworks might want to provide their own plugins. They can do so
by publishing a package `can-ssr-plugin-FOO` where FOO is the name of
the plugin that will be provided in options.

To use a plugin include a second argument of options with an array of
plugins to include. For example:

```js
var ssr = require("can-ssr");

var render = ssr({}, {
	plugins: ["foo"]
});
```

Would include the `can-ssr-plugin-foo` plugin.

If you would like to exclude all plugins, including the canjs plugin,
   make `plugins: false` the option.

Closes #87
