+++ lib/extension.js
@@ -2,85 +2,14 @@
 
 // Imports
 var loader = require("@loader");
-var stache = require("can/view/stache/stache");
-require("can/view/import/import");
+var helpers = require("@ssr/bundles");
 
-var bundles = {"@global": {}};
-var parentMap = {};
-
-function setAsBundle(name, parentName){
-	return loader.normalize(name, parentName).then(function(name) {
-		if(!bundles[name]) {
-			bundles[name] = {};
-		}
-	});
-}
-
-function isProduction(options){
-	var prod = process.env.NODE_ENV === "production";
-	if(prod) {
-		return options.fn(this);
-	} else {
-		return options.inverse(this);
-	}
-}
-
-function findBundleName(moduleName) {
-	var parent = parentMap[moduleName],
-		bundleName = parent;
-	while(parent) {
-		parent = parentMap[parent];
-		if(parent) {
-			bundleName = parent;
-		}
-	}
-	return bundleName;
-}
-
-function findBundle(moduleName){
-	var bundleName = findBundleName(moduleName);
-	return bundles[bundleName];
-}
-
-function assetHelper(type){
-	var state = this;
-	var assets = this.attr("__renderingAssets");
-	var complete = this.attr("__renderingComplete");
-	var frag = document.createDocumentFragment();
-
-	if(complete) {
-		var inserted = {};
-		assets.each(function(moduleName){
-			var bundle = findBundle(moduleName) || bundles[moduleName];
-
-			if(bundle) {
-				Object.keys(bundle).forEach(function(childName){
-					var asset = bundle[childName];
-					if(asset && asset.type === type && !inserted[asset.id]) {
-						inserted[asset.id] = true;
-						var node = asset.value.call(state);
-						node.setAttribute("asset-id", asset.id);
-						frag.appendChild(node);
-					}
-				});
-			}
-		});
-
-		var globals = bundles["@global"];
-		Object.keys(globals).forEach(function(moduleName){
-			var asset = globals[moduleName];
-			if(asset.type === type) {
-				var node = asset.value.call(state);
-				if(node.setAttribute) {
-					node.setAttribute("asset-id", "@" + asset.type);
-				}
-				frag.appendChild(node);
-			}
-		});
-	}
-
-	return frag;
-}
+// Bundle helpers
+var setAsBundle = helpers.setAsBundle;
+var findBundle = helpers.findBundle;
+var findBundleName = helpers.findBundleName;
+var bundles = helpers.bundles;
+var parentMap = helpers.parentMap;
 
 function assetRegister(moduleName, type, makeAsset) {
 	if(arguments.length === 2) {
@@ -128,9 +57,6 @@ loader.set("asset-register", loader.newModule({
 	"default": assetRegister
 }));
 
-stache.registerHelper("asset", assetHelper);
-stache.registerHelper("isProduction", isProduction);
-
 var loaderImport = loader.import;
 loader.import = function(name, options){
 	var loader = this, args = arguments;
@@ -151,24 +77,6 @@ loader.normalize = function(name, parentName){
 		}
 		return normalizedName;
 	});
-};
-
-var canImport = can.view.callbacks._tags["can-import"];
-can.view.callbacks._tags["can-import"] = function(el, tagData){
-	var root = tagData.scope.attr("%root") || tagData.scope.attr("@root");
-	var moduleName = el.getAttribute("from");
-	var templateModule = tagData.options.attr("helpers.module");
-	var parentName = templateModule ? templateModule.id : undefined;
-
-	var isAPage = !!tagData.subtemplate;
-	loader.normalize(moduleName, parentName).then(function(name){
-		if(isAPage) {
-			parentMap[name] = false;
-		}
-		root.attr("__renderingAssets").push(name);
-	});
-
-	canImport.apply(this, arguments);
 };
 
 [
