+++ lib/index.js
@@ -6,7 +6,8 @@ var makeMap = require("./make_map");
 var mergeResponseData = require("./merge_response_data");
 var wait = require("can-wait");
 
-module.exports = function(cfg){
+module.exports = function(cfg, options){
+	options = getOptions(options);
 	var steal = Steal.clone();
 	var loader = global.System = steal.System;
 
@@ -18,7 +19,7 @@ module.exports = function(cfg){
 	steal.config(cfg || {});
 
 	// Ensure the extension is loaded before the main.
-	loadExtension(loader);
+	loadExtension(loader, options);
 
 	function getAutorender(autorender){
 		// startup returns an Array in dev
@@ -137,3 +138,18 @@ module.exports = function(cfg){
 		});
 	};
 };
+
+function getOptions(opts){
+	opts = opts || {};
+
+	if(opts.plugins !== false && !opts.plugins) {
+		opts.plugins = [
+			"canjs"
+		];
+	}
+	if(!opts.plugins) {
+		opts.plugins = [];
+	}
+
+	return opts;
+}
