+++ lib/load_extension.js
@@ -28,13 +28,16 @@ function overrideConfig(loader){
 	};
 }
 
-module.exports = function(loader){
+module.exports = function(loader, options){
 	// Configure the loader to set our overrides.
 	loader.config({
 		paths: {
 			"@ssr": "file:" + path.resolve(path.join(__dirname, "extension.js")),
 			"@ssr/*": "file:" + path.resolve(path.join(__dirname, "/*.js")),
 			"@ssr/system-config": "file:" + path.resolve(path.join(__dirname, "system_config.js"))
+		},
+		map: {
+			"can-ssr-plugin-canjs": "@ssr/plugins/canjs"
 		},
 		configDependencies: [
 			"@ssr/system-config"
@@ -65,10 +68,22 @@ module.exports = function(loader){
 			}
 
 
-			return loader.import("@ssr").then(function(){
-				return loaderImport.apply(loader, args);
-			});
+			return loader.import("@ssr")
+				.then(function(){
+					return Promise.all(options.plugins.map(function(name){
+						return loader.import(pluginName(name), {
+							name: loader.main
+						});
+					}));
+				})
+				.then(function(){
+					return loaderImport.apply(loader, args);
+				});
 		}
 		return loaderImport.apply(this, arguments);
 	};
 };
+
+function pluginName(name){
+	return "can-ssr-plugin-" + name;
+}
