+++ test/tests/package.json
@@ -4,7 +4,8 @@
   "main": "progressive/index.stache!done-autorender",
   "dependencies": {
     "can": "2.3.0-pre.1",
-    "done-autorender": "0.0.7"
+    "done-autorender": "0.0.7",
+    "can-ssr-plugin-foo": "1.0.0"
   },
   "system": {
     "paths": { "./app-map": "../../app-map.js" },
